svn path=/branches/libavg_dscamera/; revision=5814

+++ src/graphics/GLTexture.cpp
@@ -29,7 +29,8 @@ GLTexture::GLTexture(const IntPoint& size, PixelFormat pf, bool bMipmap,
         unsigned wrapSMode, unsigned wrapTMode)
     : m_Size(size),
       m_pf(pf),
-      m_bMipmap(bMipmap)
+      m_bMipmap(bMipmap),
+      m_bDeleteTex(true)
 {
     if (getGLType(m_pf) == GL_FLOAT && !isFloatFormatSupported()) {
         throw Exception(AVG_ERR_UNSUPPORTED, 
@@ -53,14 +54,25 @@ GLTexture::GLTexture(const IntPoint& size, PixelFormat pf, bool bMipmap,
     glTexImage2D(GL_TEXTURE_2D, 0, getGLInternalFormat(), m_Size.x, m_Size.y, 0,
             getGLFormat(m_pf), getGLType(m_pf), 0);
     OGLErrorCheck(AVG_ERR_VIDEO_GENERAL, "GLTexture: glTexImage2D()");
+}
 
+GLTexture::GLTexture(unsigned glTexID, const IntPoint& size, PixelFormat pf, bool bMipmap,
+        bool bDeleteTex)
+    : m_Size(size),
+      m_pf(pf),
+      m_bMipmap(bMipmap),
+      m_bDeleteTex(bDeleteTex),
+      m_TexID(glTexID)
+{
 }
 
 GLTexture::~GLTexture()
 {
-    glBindTexture(GL_TEXTURE_2D, 0);
-    glDeleteTextures(1, &m_TexID);
-    OGLErrorCheck(AVG_ERR_VIDEO_GENERAL, "GLTexture: DeleteTextures()");
+    if (m_bDeleteTex) {
+        glBindTexture(GL_TEXTURE_2D, 0);
+        glDeleteTextures(1, &m_TexID);
+        OGLErrorCheck(AVG_ERR_VIDEO_GENERAL, "GLTexture: DeleteTextures()");
+    }
 }
 
 void GLTexture::activate(int textureUnit)
@@ -77,6 +89,13 @@ void GLTexture::generateMipmaps()
     activate();
     glproc::GenerateMipmap(GL_TEXTURE_2D);
     OGLErrorCheck(AVG_ERR_VIDEO_GENERAL, "GLTexture::generateMipmaps()");
+}
+
+void GLTexture::setWrapMode(unsigned wrapSMode, unsigned wrapTMode)
+{
+    activate();
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, wrapSMode);
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, wrapTMode);
 }
 
 const IntPoint& GLTexture::getSize() const
